Re-work depth bindings, remove class variables for configuration
Turns out it is easier to just use the `buffer` object passed into
the BufferTags object, instead of passing it on as variables onto
a number of functions.  The Tag object still has depth passed in
so it can remain isolated and using bindings for block tags.

Configuration also had some issues later on trying to use
class_variable_get, so I moved it to a simpler instance variable
setup.